+++ src/main/java/com/studville/studville/StudService/ServiceInterface/StudInterface.java
@@ -1,7 +1,7 @@
 package com.studville.studville.StudService.ServiceInterface;
 
 import java.util.List;
-import java.util.Optional;
+
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,9 +13,9 @@ public interface StudInterface {
     Stud addStud(Stud stud);
     List<Stud> getAllStuds();
     Page<Stud> listAllstuds(Pageable pageable);
-    Stud findStudById(Long Id);
+    Stud findStudById(Integer Id);
     Stud findStudByEmail(String email);
-    Stud updateStud(Stud stud, Long id);
-    void removeStud(Long id);
+    Stud updateStud(Stud stud, Integer id);
+    void removeStud(Integer id);
     
 }
